+++ contract/types/vote.ts
