+++ ui/src/types/vote.ts
